requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
matplotlib>=3.8.0
line-bot-sdk>=3.5.0
